Remove messages variable
